front/syn_metadata: move classify_fn_arg_ty + extract_dyn_trait_root + canonical_pat_name out of front/ast (Slice 2)
